add table inserts

+++ src/main/scala/Core/Application/Port/OpenDota/Model/ProPlayer.scala
@@ -1,7 +1,7 @@
 package Core.Application.Port.OpenDota.Model
 
 case class ProPlayer (
-    account_id: Double,
+    account_id: Int,
     steamid: String,
     avatar: String,
     avatarmedium: String,
@@ -10,15 +10,15 @@ case class ProPlayer (
     personaname: String,
     last_login: Option[String],
     full_history_time: Option[String],
-    cheese: Double,
+    cheese: Int,
     fh_unavailable: Option[Boolean],
     loccountrycode: Option[String],
     last_match_time: String,
     plus: Option[Boolean],
     name: String,
     country_code: String,
-    fantasy_role: Double,
-    team_id: Double,
+    fantasy_role: Int,
+    team_id: Int,
     team_name: Option[String],
     team_tag: Option[String],
     is_locked: Boolean,
